requirements.txt
mcp-sdk
aiohttp==3.9.1
python-dotenv==1.0.0
PyGithub>=2.1.1
pydantic>=2.5.2
typing-extensions>=4.8.0